% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Names of countries with available value sets}
\format{
A list with two elements: 3L and 5L.
}
\usage{
data(country_codes)
}
\description{
A list consisting of data frames with 2-letter code, full name and short name for every country, for which the relevant value set is available in the, respectively, 3L and 5L value_sets datasets.
}
\keyword{datasets}
