% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxwr.R
\name{.EQxwrprob}
\alias{.EQxwrprob}
\title{.EQxwrprob}
\usage{
.EQxwrprob(par = NULL)
}
\arguments{
\item{par}{Matrix of model parameters}
}
\value{
An 243 * 25 matrix with probabilities for state level transitions.
}
\description{
Takes a matrix of parameters for reverse crosswalk model, returns 243 x 25 matrix of state/level transition probabilities.
}
