% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.get_country_code}
\alias{.get_country_code}
\title{Get country code based on input string}
\usage{
.get_country_code(country, eq5d_version)
}
\arguments{
\item{country}{A character string representing the name of the country.}

\item{eq5d_version}{A character string specifying the version of EQ-5D, i.e. 3L or 5L.}
}
\value{
A character string representing the calculated country code, or `NA` if the input country name is not found in the country_codes dataset.
}
\description{
This function returns the country code based on the input string and a specified version of EQ-5D. The country code is the short name extracted from the respective country_codes dataset.
}
\examples{
.get_country_code("United Kingdom", "3L") # returns `United Kingdom`.
.get_country_code("GB", "3L") # returns `United Kingdom`.
.get_country_code("GB", "5L") 
# returns `NA` as the 5L value set is currently not available for the United Kingdom
}
