% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.get_names}
\alias{.get_names}
\title{Replace NULL names with default values}
\usage{
.get_names(df = NULL, ...)
}
\arguments{
\item{df}{a data frame; only used/supplied if levels_fu needs to be defined}

\item{...}{a list of parameters consisting of any/all of `names_eq5d`, `name_fu`, `levels_fu`, `eq5d_version`, and `name_vas`.}
}
\value{
a list of parameters with null entries replaced with default values.
}
\description{
This function takes in a list of parameters, which would be column names of the input data frame, and checks if they are null. Any nulls are replaced with default values, and the updated list of parameters is returned.
}
\examples{
.get_names(names_eq5d = c("mo", "sc", "ua", "pd", "ad"))
.get_names(names_eq5d = NULL, eq5d_version = NULL, name_vas = NULL)
.get_names(df = example_data, name_fu = NULL, levels_fu = NULL)
}
