% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.modify_ggplot_theme}
\alias{.modify_ggplot_theme}
\title{Modify ggplot2 theme}
\usage{
.modify_ggplot_theme(p)
}
\arguments{
\item{p}{ggplot2 plot}
}
\value{
ggplot2 plot with modified theme
}
\description{
Modify ggplot2 theme
}
