% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.pchc}
\alias{.pchc}
\title{Wrapper to determine Paretian Classification of Health Change (PCHC)}
\usage{
.pchc(df, group_by, add_noprobs = FALSE)
}
\arguments{
\item{df}{A data frame with EQ-5D profiles and variables to group by. The dataset is assumed to be have been ordered correctly.}

\item{group_by}{Character vector of variables to group by}

\item{add_noprobs}{Logical value indicating whether to include a separate classification for those without problems (default is FALSE)}
}
\value{
A data frame with PCHC value for each combination of the grouping variables. 
If 'add_noprobs' is TRUE, a separate classification for those without problems is also included.
}
\description{
This internal function determines Paretian Classification of Health Change (PCHC) for each combination of the variables specified in the `group_by` argument. 
It is used in the code for table_2_4-table_2_5 and figure_2_1-figure_2_4. 
An EQ-5D health state is deemed to be `better` than another if it is better on at least one dimension and is no worse on any other dimension.
An EQ-5D health state is deemed to be `worse` than another if it is worse in at least one dimension and is no better in any other dimension.
}
\examples{
df <- data.frame(id = c(1, 1, 2, 2),
                 mobility = c(1, 1, 1, 1),
                 selfcare = c(1, 1, 5, 1),
                 usualact = c(1, 1, 4, 3),
                 paindisc = c(1, 1, 1, 3),
                 anxietyd = c(1, 1, 1, 1))
.pchc(df, c("id"), add_noprobs = TRUE)
}
