% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxwr.R
\name{.pstate3t5}
\alias{.pstate3t5}
\title{.pstate3t5}
\usage{
.pstate3t5(PPP)
}
\arguments{
\item{PPP}{N x 25 matrix with probabilities per level/dimension created by EQrxwprobs}
}
\value{
An N * 3125 matrix with probabilities per state
}
\description{
Takes a N x 25 matrix with probabilities per level/dimension, and creates an N * 3125 matrix with probabilities per state
}
