% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxw.R
\name{.pstate5t3}
\alias{.pstate5t3}
\title{.pstate5t3}
\usage{
.pstate5t3(probs = .EQxwprob)
}
\arguments{
\item{probs}{15 x 5 matrix with probabilities per level/dimension, typically saved in .EQxwprob}
}
\value{
An 3125x243 matrix with probabilities per state
}
\description{
Takes a 15 x 5 matrix with probabilities per level/dimension, and creates an 3125x243 matrix with probabilities per state
}
