% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_dimension_change_table}
\alias{eq5d_profile_dimension_change_table}
\title{eq5d_profile_dimension_change_table: Changes in levels in each dimension, percentages of total and of type of change}
\usage{
eq5d_profile_dimension_change_table(
  df,
  name_id,
  names_eq5d = NULL,
  name_fu = NULL,
  levels_fu = NULL
)
}
\arguments{
\item{df}{Data frame with the EQ-5D, id and follow-up columns}

\item{name_id}{Character string for the patient id column}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column. 
If NULL (default value), the levels will be ordered in the order of appearance in df.}
}
\value{
Summary data frame
}
\description{
eq5d_profile_dimension_change_table: Changes in levels in each dimension, percentages of total and of type of change
}
\examples{
eq5d_profile_dimension_change_table(
  df = example_data,
  name_id = "id",
  names_eq5d = c("mo", "sc", "ua", "pd", "ad"),
  name_fu = "time",
  levels_fu = c("Pre-op" , "Post-op")
)
}
