% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_level_summary}
\alias{eq5d_profile_level_summary}
\title{eq5d_profile_level_summary: Frequency of levels by dimensions, cross-sectional}
\usage{
eq5d_profile_level_summary(df, names_eq5d = NULL, eq5d_version = NULL)
}
\arguments{
\item{df}{Data frame with the EQ-5D and follow-up columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}
}
\value{
Summary data frame.
}
\description{
eq5d_profile_level_summary: Frequency of levels by dimensions, cross-sectional
}
\examples{
eq5d_profile_level_summary(
 df = example_data[example_data$time == "Pre-op",],
 names_eq5d = c("mo", "sc", "ua", "pd", "ad"),
 eq5d_version = "3L"
)
}
