% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_level_summary_by_group}
\alias{eq5d_profile_level_summary_by_group}
\title{eq5d_profile_level_summary_by_group: Frequency of levels by dimensions, separated by category}
\usage{
eq5d_profile_level_summary_by_group(
  df,
  names_eq5d = NULL,
  name_cat = NULL,
  levels_cat = NULL,
  eq5d_version = NULL
)
}
\arguments{
\item{df}{Data frame with the EQ-5D and follow-up columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_cat}{Character string for the category  column. If NULL, no grouping is used, and the table reports for the total population, i.e. equal to table 1.1.1.}

\item{levels_cat}{Character vector containing the order of the values in the category column, if the wish is to have these presented in a particular order. 
If NULL (default value), unless the variable is a factor, the levels will be ordered in the order of appearance in df.}

\item{eq5d_version}{Version of the EQ-5D instrument}
}
\value{
Summary data frame.
}
\description{
eq5d_profile_level_summary_by_group: Frequency of levels by dimensions, separated by category
}
\examples{
eq5d_profile_level_summary_by_group(
 df = example_data[example_data$time == "Pre-op",],
 names_eq5d = c("mo", "sc", "ua", "pd", "ad"),
 name_cat = "procedure",
 levels_cat = c("Hip Replacement", "Knee Replacement"),
 eq5d_version = "3L"
)
}
