% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_lfs_distribution}
\alias{eq5d_profile_lfs_distribution}
\title{eq5d_profile_lfs_distribution: Distribution of the EQ-5D states by LFS (Level Frequency Score)}
\usage{
eq5d_profile_lfs_distribution(df, names_eq5d = NULL, eq5d_version = NULL)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}
}
\value{
Summary data frame
}
\description{
eq5d_profile_lfs_distribution: Distribution of the EQ-5D states by LFS (Level Frequency Score)
}
\examples{
eq5d_profile_lfs_distribution(
  example_data, 
  names_eq5d = c("mo", "sc", "ua", "pd", "ad"), 
  eq5d_version = "3L"
)
}
