% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_lfs_utility_plot}
\alias{eq5d_profile_lfs_utility_plot}
\title{eq5d_profile_lfs_utility_plot: EQ-5D values plotted against LFS}
\usage{
eq5d_profile_lfs_utility_plot(
  df,
  names_eq5d = NULL,
  eq5d_version = NULL,
  country
)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{country}{A character string representing the name of the country. 
This could be in a 2-letter format, full name or short name, as specified in the country_codes datasets.}
}
\value{
Summary plot and data used for plotting
}
\description{
eq5d_profile_lfs_utility_plot: EQ-5D values plotted against LFS
}
\examples{
tmp <- eq5d_profile_lfs_utility_plot(
 example_data, 
 names_eq5d = c("mo", "sc", "ua", "pd", "ad"), 
 eq5d_version = "3L",
 country = "UK"
)
tmp$p
tmp$plot_data
}
