% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_pchc_with_no_problems_table}
\alias{eq5d_profile_pchc_with_no_problems_table}
\title{eq5d_profile_pchc_with_no_problems_table: Changes in health according to the PCHC, taking account of those with no problems}
\usage{
eq5d_profile_pchc_with_no_problems_table(
  df,
  name_id,
  name_groupvar = NULL,
  names_eq5d = NULL,
  name_fu = NULL,
  levels_fu = NULL
)
}
\arguments{
\item{df}{Data frame with the EQ-5D, grouping, id and follow-up columns}

\item{name_id}{Character string for the patient id column}

\item{name_groupvar}{Character string for the grouping column. If NULL (default), the analysis is performed on the full population.}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column.
If NULL (default value), the levels will be ordered in the order of appearance in df.}
}
\value{
Summary data frame
}
\description{
eq5d_profile_pchc_with_no_problems_table: Changes in health according to the PCHC, taking account of those with no problems
}
\examples{
eq5d_profile_pchc_with_no_problems_table(
  df = example_data,
  name_id = "id",
  name_groupvar = "procedure",
  names_eq5d = c("mo", "sc", "ua", "pd", "ad"),
  name_fu = "time",
  levels_fu = c("Pre-op" , "Post-op")
)
}
