% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_profile_top_states}
\alias{eq5d_profile_top_states}
\title{eq5d_profile_top_states: Prevalence of the 10 most frequently observed self-reported health states}
\usage{
eq5d_profile_top_states(df, names_eq5d = NULL, eq5d_version = NULL, n = 10)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{n}{Number of most frequently observed states to display (default 10)}
}
\value{
Summary data frame
}
\description{
eq5d_profile_top_states: Prevalence of the 10 most frequently observed self-reported health states
}
\examples{
eq5d_profile_top_states(
  df = example_data[example_data$time == "Pre-op",],
  names_eq5d = c("mo", "sc", "ua", "pd", "ad"),
  eq5d_version = "3L",
  n = 10
)
}
