% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_utility_over_time_plot}
\alias{eq5d_utility_over_time_plot}
\title{eq5d_utility_over_time_plot: EQ-5D values by timepoints: mean values and 95\% confidence intervals}
\usage{
eq5d_utility_over_time_plot(
  df,
  names_eq5d = NULL,
  name_fu = NULL,
  levels_fu = NULL,
  eq5d_version = NULL,
  country
)
}
\arguments{
\item{df}{Data frame with the VAS columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column. 
If NULL (default value), the levels will be ordered in the order of appearance in df.}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{country}{A character string representing the name of the country. 
This could be in a 2-letter format, full name or short name, as specified in the country_codes datasets.}
}
\value{
Summary plot and data used for plotting
}
\description{
eq5d_utility_over_time_plot: EQ-5D values by timepoints: mean values and 95\% confidence intervals
}
\examples{
tmp <- eq5d_utility_over_time_plot(
 example_data,
 names_eq5d = c("mo", "sc", "ua", "pd", "ad"), 
 name_fu = "time",
 levels_fu = c('Pre-op', 'Post-op'),
 eq5d_version = "3L",
 country = "UK"
)
tmp$p
tmp$plot_data
}
