% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_utility_summary_by_group}
\alias{eq5d_utility_summary_by_group}
\title{eq5d_utility_summary_by_group:EQ-5D values: by groupvar}
\usage{
eq5d_utility_summary_by_group(
  df,
  names_eq5d = NULL,
  name_groupvar = NULL,
  eq5d_version = NULL,
  country
)
}
\arguments{
\item{df}{Data frame with the EQ-5D, follow-up and grouping columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_groupvar}{Character string for the grouping column. If NULL (default), the analysis is performed on the full population.}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{country}{A character string representing the name of the country.}
}
\value{
Summary data frame
}
\description{
eq5d_utility_summary_by_group:EQ-5D values: by groupvar
}
\examples{
eq5d_utility_summary_by_group(
  example_data,
  names_eq5d = c("mo", "sc", "ua", "pd", "ad"),
  name_groupvar = "procedure",
  eq5d_version = "3L",
  country = "UK"
)
}
