% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_vas_distribution_table}
\alias{eq5d_vas_distribution_table}
\title{eq5d_vas_distribution_table: EQ VAS Scores frequency of mid-points}
\usage{
eq5d_vas_distribution_table(df, name_vas = NULL, add_na_total = TRUE)
}
\arguments{
\item{df}{Data frame with the VAS column}

\item{name_vas}{Character string for the VAS column}

\item{add_na_total}{Logical, whether to add summary of the missing, and across the Total, data}
}
\value{
Summary data frame
}
\description{
eq5d_vas_distribution_table: EQ VAS Scores frequency of mid-points
}
\examples{
eq5d_vas_distribution_table(
  example_data,
  name_vas = 'vas', 
  add_na_total =  TRUE
  )
}
