% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_vas_grouped_distribution_plot}
\alias{eq5d_vas_grouped_distribution_plot}
\title{eq5d_vas_grouped_distribution_plot: Mid-point EQ VAS scores}
\usage{
eq5d_vas_grouped_distribution_plot(df, name_vas = NULL)
}
\arguments{
\item{df}{Data frame with the VAS column}

\item{name_vas}{Character string for the VAS column}
}
\value{
Summary plot and data used for plotting
}
\description{
eq5d_vas_grouped_distribution_plot: Mid-point EQ VAS scores
}
\examples{
tmp <- eq5d_vas_grouped_distribution_plot(example_data, name_vas = 'vas')
tmp$p
tmp$plot_data
}
