% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{eq5d_vas_summary}
\alias{eq5d_vas_summary}
\title{eq5d_vas_summary: EQ VAS Score by timepoints}
\usage{
eq5d_vas_summary(df, name_vas = NULL, name_fu = NULL, levels_fu = NULL)
}
\arguments{
\item{df}{Data frame with the VAS and the follow-up columns}

\item{name_vas}{Character string for the VAS column}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column.}
}
\value{
Summary data frame
}
\description{
eq5d_vas_summary: EQ VAS Score by timepoints
}
\examples{
eq5d_vas_summary(
  example_data,
  name_vas = 'vas', 
  name_fu = 'time', 
  levels_fu = c('Pre-op', 'Post-op')
)
}
