% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R
\name{eqvs_load}
\alias{eqvs_load}
\title{eqvs_load}
\usage{
eqvs_load(loadPath)
}
\arguments{
\item{loadPath}{The path from which to load the cache data.}
}
\value{
TRUE if loading is successful, FALSE otherwise.
}
\description{
Load cache data from a specified path.
}
