% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R, R/eqxw.R
\name{eqxw}
\alias{eqxw}
\title{eqxw}
\usage{
eqxw(x, country = NULL, dim.names = c("mo", "sc", "ua", "pd", "ad"))

eqxw(x, country = NULL, dim.names = c("mo", "sc", "ua", "pd", "ad"))
}
\arguments{
\item{x}{A vector of 5-digit EQ-5D-5L state indexes or a matrix/data.frame with columns corresponding to EQ-5D state dimensions}

\item{country}{String vector indicating country name(s) or country code(s) according to ISO3166 Alpha 2 or Alpha 3}

\item{dim.names}{A vector of dimension names to identify dimension columns}
}
\value{
A vector of crosswalk EQ-5D-3L values or data.frame with one column for each value set requested.

A vector of reverse crosswalk values or data.frame with one column per reverse crosswalk set requested.
}
\description{
Get EQ-5D-3L crosswalk index values from individual responses to the five dimensions of the EQ-5D-5L (-3L value set applied to -5L health states).

Get crosswalk values
}
\examples{
eqxw(c(11111, 12321, 32423, 55555), 'US') # US -3L crosswalk value set
eqxw(make_all_EQ_states('5L'), c('DK', 'CA')) # Danish and Canada -3L crosswalk value sets 
eqxw(c(11111, 12521, 32123, 55555), 'US')
eqxw(make_all_EQ_states('5L'), c('DK', 'US'))
}
