% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxw.R
\name{eqxw_NICE}
\alias{eqxw_NICE}
\title{eqxw_NICE}
\usage{
eqxw_NICE(
  x,
  age,
  male,
  dim.names = c("mo", "sc", "ua", "pd", "ad"),
  bwidth = 0
)
}
\arguments{
\item{x}{A vector of 5-digit EQ-5D-5L states (domain scores) or a summary score.}

\item{age}{A numeric vector or column name indicating respondent age.}

\item{male}{A numeric vector (1=male, 0=female) or column name indicating gender.}

\item{dim.names}{A vector of dimension names for EQ-5D states (default: c("mo", "sc", "ua", "pd", "ad")).}

\item{bwidth}{Numeric. Bandwidth for kernel smoothing when using summary scores.}
}
\value{
A vector or data frame with crosswalked EQ-5D-3L utilities.
}
\description{
Crosswalks EQ-5D-5L responses to EQ-5D-3L utilities using NICE's mapping.
}
\examples{
eqxw_NICE(c(11111, 12321, 32123, 33333), age = c(30, 40, 55, 70), male = c(1, 0, 1, 0))
eqxw_NICE(df, age = "age", male = "male")
}
