% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R, R/eqxwr.R
\name{eqxwr}
\alias{eqxwr}
\title{eqxwr}
\usage{
eqxwr(x, country = NULL, dim.names = c("mo", "sc", "ua", "pd", "ad"))

eqxwr(x, country = NULL, dim.names = c("mo", "sc", "ua", "pd", "ad"))
}
\arguments{
\item{x}{A vector of 5-digit EQ-5D-3L state indexes or a matrix/data.frame with columns corresponding to EQ-5D state dimensions}

\item{country}{String vector indicating country name(s) or country code(s) according to ISO3166 Alpha 2 or Alpha 3}

\item{dim.names}{A vector of dimension names to identify dimension columns}
}
\value{
A vector of EQ-5D-5L values or data.frame with one column for each value set requested.

A vector of reverse crosswalk values or data.frame with one column per reverse crosswalk set requested.
}
\description{
Get EQ-5D-5L crosswalk reverse index values from individual responses to the five dimensions of the EQ-5D-3L (-5L value set applied to -3L health states).

Get reverse crosswalk values
}
\examples{
eqxwr(c(11111, 12321, 32123, 33333), 'IT') # Italy -5L crosswalk reverse value set
eqxwr(make_all_EQ_states('3L'), c('Japan', 'China')) # Japon and China -5L crosswalk reverse value sets 
eqxwr(c(11111, 12321, 32123, 33333), 'US')
eqxwr(make_all_EQ_states('3L'), c('DK', 'US'))
}
