% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxwr.R
\name{eqxwr_add}
\alias{eqxwr_add}
\title{eqxwr_add}
\usage{
eqxwr_add(
  df,
  version = "5L",
  country = NULL,
  description = NULL,
  code2L = NULL,
  code3L = NULL,
  three_letter_code = NULL
)
}
\arguments{
\item{df}{A data.frame or file name pointing to csv file. The contents of the data.frame or csv file should be exactly two columns: state, containing a list of all 3125 (for 5L) or 243 (for 3L) EQ-5D health state vectors, and a column of corresponding utility values, with a suitable name.}

\item{version}{Version of the EQ-5D instrument. Can take values 5L (default) or 3L.}

\item{country}{Optional string. If not NULL, will be used as a country description for the user-defined value set.}

\item{description}{Optional string. If not NULL, will be used as a descriptive text for the user-defined value set.}

\item{code2L}{Optional string. If not NULL, will be used as the two-digit code for the value set. Must be different from any existing national value set code.}
}
\value{
True/False, indicating success or error.
}
\description{
Add user-defined EQ-5D value set to reverse crosswalk options.
}
\examples{
# make nonsense value set
new_df <- data.frame(state = make_all_EQ_indexes(), TEST = runif(3125))
# Add as value set for Fantasia
eqxwr_add(new_df, version = "5L", country = 'Fantasia')
}
