% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxwr.R
\name{eqxwr_display}
\alias{eqxwr_display}
\title{eqxwr_display}
\usage{
eqxwr_display(return_df = FALSE, version = "5L")
}
\arguments{
\item{return_df}{If set to TRUE, the function will return information on the names of the available value sets in a data.frame. Defaults to FALSE}

\item{version}{Version of the EQ-5D instrument. Can take values 5L (default) or 3L.}
}
\value{
Default NULL, if return_df == TRUE, returns a data.frame with the displayed information.
}
\description{
Display available reverse crosswalk value sets.
}
\examples{
# Display available value sets.
eqxwr_display
}
