% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{example_data}
\format{
A data frame with 10000 rows and 14 variables:
\describe{
  \item{\code{id}}{double Patient id}
  \item{\code{time}}{character Follow-up (Pre-op / Post-op)}
  \item{\code{mo}}{double EQ-5D-5L Mobility dimension}
  \item{\code{sc}}{double EQ-5D-5L Self-care dimension}
  \item{\code{ua}}{double EQ-5D-5L Usual activities dimension}
  \item{\code{pd}}{double EQ-5D-5L Pain / discomfort dimension}
  \item{\code{ad}}{double EQ-5D-5L Anxiety/depression dimension}
  \item{\code{vas}}{double Value of the VAS scale measurememnt}
  \item{\code{providercode}}{character Provider code}
  \item{\code{procedure}}{character Type of surgery}
  \item{\code{year}}{character Year of intervention}
  \item{\code{ageband}}{character Age in pre-defined ranges} 
  \item{\code{gender}}{character Patient's gender (Female / Male)}
}
}
\usage{
data(example_data)
}
\description{
A dataset containing patient-level data in a long format.
}
\keyword{datasets}
