% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_value_sets.R
\name{fetch_available_value_sets}
\alias{fetch_available_value_sets}
\title{Fetch the index of available value sets from GitHub}
\usage{
fetch_available_value_sets(version)
}
\arguments{
\item{version}{Character. One of "3L", "5L", or "Y3L".}
}
\value{
A data frame with columns Version, Name, Name_short,
  Country_code, VS_code, doi. Returns NULL if fetch fails.
}
\description{
Fetch the index of available value sets from GitHub
}
\keyword{internal}
