% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{figure_1_3_1}
\alias{figure_1_3_1}
\title{Figure 1.3.1: EQ-5D values plotted against LSS}
\usage{
figure_1_3_1(df, names_eq5d = NULL, eq5d_version = NULL, country)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{country}{A character string representing the name of the country. 
This could be in a 2-letter format, full name or short name, as specified in the country_codes datasets.}
}
\value{
Summary plot and data used for plotting
}
\description{
Figure 1.3.1: EQ-5D values plotted against LSS
}
\examples{
tmp <- figure_1_3_1(df = example_data, country = "USA")
tmp$p
tmp$plot_data
}
