% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{figure_2_1}
\alias{figure_2_1}
\title{Figure 2.1: EQ VAS scores}
\usage{
figure_2_1(df, name_vas = NULL)
}
\arguments{
\item{df}{Data frame with the VAS column}

\item{name_vas}{Character string for the VAS column}
}
\value{
Summary plot and data used for plotting
}
\description{
Figure 2.1: EQ VAS scores
}
\examples{
tmp <- figure_2_1(df = example_data)
tmp$p
tmp$plot_data
}
