% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{figure_3_2}
\alias{figure_3_2}
\title{Figure 3.2: Mid-point EQ VAS scores}
\usage{
figure_3_2(df, name_vas = NULL)
}
\arguments{
\item{df}{Data frame with the VAS column}

\item{name_vas}{Character string for the VAS column}
}
\value{
Summary plot and data used for plotting
}
\description{
Figure 3.2: Mid-point EQ VAS scores
}
\examples{
tmp <- figure_3_2(df = example_data)
tmp$p
tmp$plot_data
}
