% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{figure_3_5}
\alias{figure_3_5}
\title{Figure 3.5: EQ-5D values: smoothed lines and confidence intervals by groupvar}
\usage{
figure_3_5(
  df,
  names_eq5d = NULL,
  name_vas = NULL,
  eq5d_version = NULL,
  country
)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_vas}{Character string for the VAS column}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{country}{A character string representing the name of the country. 
This could be in a 2-letter format, full name or short name, as specified in the country_codes datasets.}
}
\value{
Summary plot and data used for plotting
}
\description{
Figure 3.5: EQ-5D values: smoothed lines and confidence intervals by groupvar
}
\examples{
tmp <- figure_3_5(df = example_data, country = "USA")
tmp$p
tmp$plot_data
}
