% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_dir}
\alias{get_cache_dir}
\title{Get the eq5dsuite cache directory}
\usage{
get_cache_dir()
}
\value{
A character string giving the path to the cache directory.
}
\description{
Returns the path to the user-specific cache directory used by
eq5dsuite to store the date of the last value set check.
The directory is created if it does not exist.
}
\keyword{internal}
