% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_last_checked}
\alias{get_last_checked}
\title{Get the date of the last value set check}
\usage{
get_last_checked()
}
\value{
A \code{Date} object.
}
\description{
Reads the cached date of the last value set update check from the
user cache directory. Returns \code{1970-01-01} if no cache file
exists (i.e., the check has never been run).
}
\keyword{internal}
