% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_value_sets.R
\name{install_value_set}
\alias{install_value_set}
\title{Download and install a single value set}
\usage{
install_value_set(vs_code, version, meta_row)
}
\arguments{
\item{vs_code}{Character. The VS_code to download.}

\item{version}{Character. One of "3L", "5L", or "Y3L".}

\item{meta_row}{A single-row data frame from the index CSV.}
}
\value{
Logical. TRUE if successful, FALSE otherwise.
}
\description{
Download and install a single value set
}
\keyword{internal}
