% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{is_update_due}
\alias{is_update_due}
\title{Check whether a value set update check is due}
\usage{
is_update_due(threshold_days = 60)
}
\arguments{
\item{threshold_days}{A single positive integer giving the number of
days after which an update check is considered overdue. Defaults
to \code{60}.}
}
\value{
A single logical value.
}
\description{
Compares the date of the last value set check (from the user cache)
against \code{Sys.Date()}. Returns \code{TRUE} if more than
\code{threshold_days} days have elapsed, indicating that the user
should be prompted to run \code{update_value_sets()}.
}
\keyword{internal}
