% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrations.R
\name{set_applied_migrations}
\alias{set_applied_migrations}
\title{Save the list of applied migrations to cache}
\usage{
set_applied_migrations(migrations)
}
\arguments{
\item{migrations}{Character vector of migration IDs to save.}
}
\value{
Invisibly returns the saved migrations.
}
\description{
Save the list of applied migrations to cache
}
\keyword{internal}
