% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_last_checked}
\alias{set_last_checked}
\title{Set the date of the last value set check}
\usage{
set_last_checked(date = Sys.Date())
}
\arguments{
\item{date}{A \code{Date} object. Defaults to \code{Sys.Date()}.}
}
\value{
Invisibly returns \code{date}.
}
\description{
Writes the given date to the user cache directory so that future
calls to \code{get_last_checked()} and \code{is_update_due()} can
determine how long ago the check was performed.
}
\keyword{internal}
