% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_1_1_1}
\alias{table_1_1_1}
\title{Table 1.1.1: Frequency of levels by dimensions, cross-sectional}
\usage{
table_1_1_1(df, names_eq5d = NULL, eq5d_version = NULL)
}
\arguments{
\item{df}{Data frame with the EQ-5D and follow-up columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}
}
\value{
Summary data frame.
}
\description{
Table 1.1.1: Frequency of levels by dimensions, cross-sectional
}
\examples{
table_1_1_1(df = example_data)
table_1_1_1(df = example_data, eq5d_version = "3L")
}
