% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_1_1_3}
\alias{table_1_1_3}
\title{Table 1.1.3: Prevalence of the 10 most frequently observed self-reported health states}
\usage{
table_1_1_3(df, names_eq5d = NULL, eq5d_version = NULL, n = 10)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{n}{Number of most frequently observed states to display (default 10)}
}
\value{
Summary data frame
}
\description{
Table 1.1.3: Prevalence of the 10 most frequently observed self-reported health states
}
\examples{
table_1_1_3(df = example_data)
table_1_1_3(df = example_data, n = 5)
table_1_1_3(df = example_data, eq5d_version = "3L")
}
