% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_1_2_3}
\alias{table_1_2_3}
\title{Table 1.2.3: Changes in health according to the PCHC, taking account of those with no problems}
\usage{
table_1_2_3(
  df,
  name_id,
  name_groupvar,
  names_eq5d = NULL,
  name_fu = NULL,
  levels_fu = NULL
)
}
\arguments{
\item{df}{Data frame with the EQ-5D, grouping, id and follow-up columns}

\item{name_id}{Character string for the patient id column}

\item{name_groupvar}{Character string for the grouping column}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column. 
If NULL (default value), the levels will be ordered in the order of appearance in df.}
}
\value{
Summary data frame
}
\description{
Table 1.2.3: Changes in health according to the PCHC, taking account of those with no problems
}
\examples{
table_1_2_3(df = example_data, name_groupvar = "surgtype", name_id = "id")
}
