% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_1_2_4}
\alias{table_1_2_4}
\title{Table 1.2.4: Changes in levels in each dimension, percentages of total and of type of change}
\usage{
table_1_2_4(df, name_id, names_eq5d = NULL, name_fu = NULL, levels_fu = NULL)
}
\arguments{
\item{df}{Data frame with the EQ-5D, id and follow-up columns}

\item{name_id}{Character string for the patient id column}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column. 
If NULL (default value), the levels will be ordered in the order of appearance in df.}
}
\value{
Summary data frame
}
\description{
Table 1.2.4: Changes in levels in each dimension, percentages of total and of type of change
}
\examples{
table_1_2_4(df = example_data, name_id = "id")
}
