% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_1_3_4}
\alias{table_1_3_4}
\title{Table 1.3.4: Summary statistics of EQ-5D values by LFS (Level Frequency Score)}
\usage{
table_1_3_4(df, names_eq5d = NULL, eq5d_version = NULL, country)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{country}{A character string representing the name of the country. 
This could be in a 2-letter format, full name or short name, as specified in the country_codes datasets.}
}
\value{
Summary data frame
}
\description{
Table 1.3.4: Summary statistics of EQ-5D values by LFS (Level Frequency Score)
}
\examples{
table_1_3_4(df = example_data, country = "Denmark")
table_1_3_4(df = example_data, eq5d_version = "3L", country = "Denmark")
}
