% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_2_1}
\alias{table_2_1}
\title{Table 2.1: EQ VAS Score by timepoints}
\usage{
table_2_1(df, name_vas = NULL, name_fu = NULL, levels_fu = NULL)
}
\arguments{
\item{df}{Data frame with the VAS and the follow-up columns}

\item{name_vas}{Character string for the VAS column}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column.}
}
\value{
Summary data frame
}
\description{
Table 2.1: EQ VAS Score by timepoints
}
\examples{
table_2_1(df = example_data)
}
