% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_2_2}
\alias{table_2_2}
\title{Table 2.2: EQ VAS Scores frequency of mid-points}
\usage{
table_2_2(df, name_vas = NULL, add_na_total = TRUE)
}
\arguments{
\item{df}{Data frame with the VAS column}

\item{name_vas}{Character string for the VAS column}

\item{add_na_total}{Logical, whether to add summary of the missing, and across the Total, data}
}
\value{
Summary data frame
}
\description{
Table 2.2: EQ VAS Scores frequency of mid-points
}
\examples{
table_2_2(df = example_data)
}
