% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{table_2_8}
\alias{table_2_8}
\title{Table 2.8: Number of observations in the LFS (Level Frequency Score) according to the EQ-5D values}
\usage{
table_2_8(df, names_eq5d = NULL, eq5d_version = NULL, country)
}
\arguments{
\item{df}{Data frame with the EQ-5D columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{eq5d_version}{Version of the EQ-5D instrument}

\item{country}{A character string representing the name of the country. 
This could be in a 2-letter format, full name or short name, as specified in the country_codes datasets.}
}
\value{
Summary data frame
}
\description{
Table 2.8: Number of observations in the LFS (Level Frequency Score) according to the EQ-5D values
}
\examples{
table_2_8(df = example_data, country = "United States")
table_2_8(df = example_data, eq5d_version = "3L", country = "United States")
}
