% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R
\name{toEQ5Ddims}
\alias{toEQ5Ddims}
\title{toEQ5Ddims}
\usage{
toEQ5Ddims(x, dim.names = c("mo", "sc", "ua", "pd", "ad"))
}
\arguments{
\item{x}{A vector of 5-digit EQ-5D state indexes.}

\item{dim.names}{A vector of dimension names to be used as names for output columns.}
}
\value{
A data.frame with 5 columns, one for each EQ-5D dimension, with names from dim.names argument.
}
\description{
Generate dimension vectors based on state index
}
\examples{
toEQ5Ddims(c(12345, 54321, 12321))
}
