% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{value_sets}
\alias{value_sets}
\title{Available 3L and 5L value sets}
\format{
A list with two elements: 3L and 5L.
}
\usage{
data(value_sets)
}
\description{
A list consisting of two data frames containing available value sets, respectively for 3L and 5L data. Column names correspond to the short names stored in the country_codes datasets.
}
\keyword{datasets}
