% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_aux.R
\name{.getmode}
\alias{.getmode}
\title{Get the mode of a vector.}
\usage{
.getmode(v)
}
\arguments{
\item{v}{A numeric or character vector.}
}
\value{
The mode of `v`.
}
\description{
This function calculates the mode of a numeric or character vector. 
If there are multiple modes, the first one is returned. 
The code is taken from an \href{https://www.tutorialspoint.com/r/r_mean_median_mode.htm}{R help page}.
}
\examples{
.getmode(c(1, 2, 3, 3))
.getmode(c("a", "b", "b", "c"))
}
