% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EQ_functions.R
\name{eqvs_display}
\alias{eqvs_display}
\title{eqvs_display}
\usage{
eqvs_display(version = "5L", return_df = FALSE)
}
\arguments{
\item{version}{Version of the EQ-5D instrument. Can take values 5L (default) or 3L.}

\item{return_df}{If set to TRUE, the function will return information on the names of the available value sets in a data.frame. Defaults to FALSE}
}
\value{
Default NULL, if return_df == TRUE, returns a data.frame with the displayed information.
}
\description{
Display available value sets, which can also be used as (reverse) crosswalks.
}
\examples{
# Display available value sets.
eqvs_display
}
