% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxw.R
\name{eqxw}
\alias{eqxw}
\title{eqxw}
\usage{
eqxw(x, country = NULL, dim.names = c("mo", "sc", "ua", "pd", "ad"))
}
\arguments{
\item{x}{A vector of 5-digit EQ-5D-5L state indexes or a matrix/data.frame with columns corresponding to EQ-5D state dimensions}

\item{country}{String vector indicating country names or  ISO3166 Alpha 2 / 3 country codes.}

\item{dim.names}{A vector of dimension names to identify dimension columns}
}
\value{
A vector of reverse crosswalk values or data.frame with one column per reverse crosswalk set requested.
}
\description{
Get crosswalk values
}
\examples{
eqxw(c(11111, 12521, 32123, 55555), 'US')
eqxw(make_all_EQ_states('5L'), c('DK', 'US'))
}
