% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqxwr.R
\name{eqxwr}
\alias{eqxwr}
\title{eqxwr}
\usage{
eqxwr(x, country = NULL, dim.names = c("mo", "sc", "ua", "pd", "ad"))
}
\arguments{
\item{x}{A vector of 5-digit EQ-5D-3L state indexes or a matrix/data.frame with columns corresponding to EQ-5D state dimensions}

\item{country}{String vector indicating country names or  ISO3166 Alpha 2 / 3 country codes.}

\item{dim.names}{A vector of dimension names to identify dimension columns}
}
\value{
A vector of reverse crosswalk values or data.frame with one column per reverse crosswalk set requested.
}
\description{
Get reverse crosswalk values
}
\examples{
eqxwr(c(11111, 12321, 32123, 33333), 'US')
eqxwr(make_all_EQ_states('3L'), c('DK', 'US'))
}
