% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{An example patient-level dataset}
\format{
A data frame with 6600 rows and 15 variables
}
\usage{
data(example_data)
}
\description{
A dataset containing patient-level data in a long format.
}
\details{
\itemize{
 \item id: patient id
 \item mobility, selfcare, usualact, paindisc, anxietyd: the five EQ-5D domains
 \item vas: value of the VAS scale measurememnt
 \item fu: follow-up (baseline / follow-up)
 \item year_range: time period for the follow-up (2009-2010 / 2010-2011 / 2011-2012); could be used as an alternative follow-up variable
 \item month: another alternative follow-up variable (1, ..., 15)
 \item surgtype: type of surgery (Cataract / Hernia / Hip / Knee / Veins)
 \item gender: patient's gender (Female / Male)
 \item age: patient's gender in years
}
}
\keyword{datasets}
