% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq5d_devlin.R
\name{figure_1_2_2}
\alias{figure_1_2_2}
\title{Figure 1.2.2: Percentage of respondents who improved overall by the dimensions (\%)}
\usage{
figure_1_2_2(df, names_eq5d = NULL, name_fu = NULL, levels_fu = NULL, name_id)
}
\arguments{
\item{df}{Data frame with the EQ-5D, follow-up and patient id columns}

\item{names_eq5d}{Character vector of column names for the EQ-5D dimensions}

\item{name_fu}{Character string for the follow-up column}

\item{levels_fu}{Character vector containing the order of the values in the follow-up column. 
If NULL (default value), the levels will be ordered in the order of appearance in df.}

\item{name_id}{Character string for the patient id column}
}
\value{
Summary plot and data used for plotting
}
\description{
Figure 1.2.2: Percentage of respondents who improved overall by the dimensions (\%)
}
\examples{
tmp <- figure_1_2_2(df = example_data, name_fu = "year_range", name_id = "id")
tmp$p
tmp$plot_data
}
